requirements.txt
## this is the original
fastapi[all]
psycopg2-binary
pydantic
redis
python-jose[cryptography]
sqlalchemy
bcrypt

## with version number compatible
# annotated-types==0.6.0
# anyio==4.3.0
# asgiref==3.7.2
# bcrypt==4.1.2
# certifi==2024.2.2
# cffi==1.16.0
# click==8.1.7
# cryptography==42.0.5
# dnspython==2.6.1
# ecdsa==0.18.0
# email-validator==2.1.0.post1
# fastapi==0.68.2
# greenlet==3.0.3
# h11==0.14.0
# httpcore==1.0.4
# httptools==0.6.1
# httpx==0.27.0
# idna==3.6
# itsdangerous==2.1.2
# Jinja2==3.1.3
# MarkupSafe==2.1.5
# orjson==3.9.15
# psycopg2-binary==2.9.9
# pyasn1==0.5.1
# pycparser==2.21
# pydantic==1.10.14
# pydantic-extra-types==2.5.0
# pydantic-settings==2.2.1
# pydantic_core==2.16.3
# python-dotenv==1.0.1
# python-jose==3.3.0
# python-multipart==0.0.9
# PyYAML==6.0.1
# redis==5.0.1
# rsa==4.9
# six==1.16.0
# sniffio==1.3.1
# SQLAlchemy==2.0.27
# starlette==0.14.2
# typing_extensions==4.10.0
# ujson==5.9.0
# uvicorn==0.15.0
# uvloop==0.19.0
# watchfiles==0.21.0
# websockets==12.0